remove table validation checks
Remove validation checks on table. The validation checks mainly fail for
debug builds and what they do is to actually validate the metadata for a
table. While dealing with indices for dictionaries, it looks like for
compatibility, we accept signed or unsigned uint32 or uint64 integers.
However, arrow dictionary builder AppendIndices only accepts signed
integers. So, we end up appending signed integers while dictionary
schema could still be unsigned integers. So, when the validation
happens, it checks if the dictionary schema and the actual column data
are the same types. Since it's possible to have uint64 schema with int64
columns, this leads to errors like (and similar error for uint32):
```
Column data for field 0 with type dictionary<values=string, indices=int64,
ordered=0> is inconsistent with schema dictionary<values=string,
indices=uint64, ordered=0>
```